+++ ytdl_api/callbacks.py
@@ -37,6 +37,7 @@ async def on_download_start_callback(
     await queue.put(
         download.client_id,
         DownloadStatusInfo(
+            title=download.title,
             client_id=download.client_id,
             media_id=download.media_id,
             status=DownloadStatus.DOWNLOADING,
@@ -56,6 +57,7 @@ async def on_pytube_progress_callback(
     Callback which will be used in Pytube's progress update callback
     """
     download_proress = DownloadStatusInfo(
+        title=download.title,
         client_id=download.client_id,
         media_id=download.media_id,
         status=DownloadStatus.DOWNLOADING,
@@ -74,6 +76,7 @@ async def on_ytdlp_progress_callback(progress: DownloadDataInfo, **kwargs):
     queue: NotificationQueue = kwargs["queue"]
     progress = extract_percentage_progress(progress.get("_percent_str"))
     download_proress = DownloadStatusInfo(
+        title=download.title,
         client_id=download.client_id,
         media_id=download.media_id,
         status=DownloadStatus.DOWNLOADING,
@@ -98,6 +101,7 @@ async def on_start_converting(
     await queue.put(
         download.client_id,
         DownloadStatusInfo(
+            title=download.title,
             client_id=download.client_id,
             media_id=download.media_id,
             status=download.status,
@@ -152,6 +156,8 @@ async def on_finish_callback(
     await queue.put(
         download.client_id,
         DownloadStatusInfo(
+            title=download.title,
+            filesize_hr=file_size_hr,
             client_id=download.client_id,
             media_id=download.media_id,
             status=status,
@@ -177,6 +183,7 @@ async def on_error_callback(
     await queue.put(
         download.client_id,
         download_progress=DownloadStatusInfo(
+            title=download.title,
             client_id=download.client_id,
             media_id=download.media_id,
             status=DownloadStatus.FAILED,