+++ CHANGELOG.md
@@ -5,11 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.4] - 2023-11-20
+### Changed
+- Update download progress info json.
+- Logging.
+### Add
+- Added volume config for app.
+
 ## [1.4.3] - 2023-11-20
 ### Changed
 - Improvement logging in `on_finish_callback` function.
 ### Fix
-- Fixed empty file size in downloads
+- Fixed empty file size in downloads.
 ### Added
 - Field with human-readable value for file size.